order-service 추가 주문 추가 기능 일부 구현

+++ cart-service/src/main/java/com/megamaker/cartservice/dto/RequestAddCart.java
@@ -10,5 +10,5 @@ import java.util.Map;
 @Getter
 public class RequestAddCart {
     private Long storeId;
-    private Map<Integer, Integer> productQuantityMap;
+    private Map<Long, Integer> productQuantityMap;
 }
